Added non linear controller. Upnext: tuning the controller parameters

+++ boom-boat.cpp
@@ -9,17 +9,17 @@
 #include "json/json.hpp"
 using json = nlohmann::json;
 
-using Eigen::Vector3f;
-using Eigen::Vector2f;
-using Eigen::Matrix3f;
-using Eigen::VectorXf;
+using Eigen::Vector3d;
+using Eigen::Vector2d;
+using Eigen::Matrix3d;
+using Eigen::VectorXd;
 
 // // These parameters are defined in the json file
-// const float DEFAULT_CAPACITY = 100;
-// const float DEFAULT_INITIAL_FUEL = DEFAULT_CAPACITY;
+// const double DEFAULT_CAPACITY = 100;
+// const double DEFAULT_INITIAL_FUEL = DEFAULT_CAPACITY;
 
 
-// const float PI = 3.141592653589793;
+// const double PI = 3.141592653589793;
 
 // Default constructor
 BoomBoat::BoomBoat() : GenericBoat() {
@@ -31,16 +31,16 @@ BoomBoat::BoomBoat() : GenericBoat() {
     // json params;
     // file >> params;
     this->load_boom_boat_params("params.json");
-    this->fuel = this->boom_boat_params["initial_fuel"].get<float>();
-    this->cap = this->boom_boat_params["waste_tank_capacity"].get<float>();
+    this->fuel = this->boom_boat_params["initial_fuel"].get<double>();
+    this->cap = this->boom_boat_params["waste_tank_capacity"].get<double>();
     this->tank_curr = 0.0;
     this->pos << 0, 0, 0;
     this->vel << 0, 0, 0;
-    this->set_control(Vector2f(0.0, 0.0));
+    this->set_control(Vector2d(0.0, 0.0));
 }
 
 // Constructor with position
-BoomBoat::BoomBoat(Vector3f pos) : GenericBoat() {
+BoomBoat::BoomBoat(Vector3d pos) : GenericBoat() {
     // // Load parameters from params.json
     // std::ifstream file("params.json");
     // if (!file.is_open()) {
@@ -49,28 +49,28 @@ BoomBoat::BoomBoat(Vector3f pos) : GenericBoat() {
     // json params;
     // file >> params;
     this->load_boom_boat_params("params.json");
-    this->fuel = this->boom_boat_params["initial_fuel"].get<float>();
-    this->cap = this->boom_boat_params["waste_tank_capacity"].get<float>();
+    this->fuel = this->boom_boat_params["initial_fuel"].get<double>();
+    this->cap = this->boom_boat_params["waste_tank_capacity"].get<double>();
     this->tank_curr = 0.0;
     this->pos = pos;
     this->vel << 0, 0, 0;
-    this->set_control(Vector2f(0.0, 0.0));
+    this->set_control(Vector2d(0.0, 0.0));
 }
 
 // Copy constructor
 BoomBoat::BoomBoat(const BoomBoat &boom_boat) : GenericBoat(boom_boat),
  fuel(boom_boat.fuel), tank_curr(boom_boat.tank_curr), cap(boom_boat.cap) {
-    this->set_control(Vector2f(0.0, 0.0));
+    this->set_control(Vector2d(0.0, 0.0));
     this->boom_boat_params = boom_boat.boom_boat_params;
  }
 
 
 // Parameterized constructor
-BoomBoat::BoomBoat(float radius, float mass, float inertia, float mu_l,
- float mu_ct, float mu_r, Vector3f pos, Vector3f vel, float fuel, float cap,
-  float F_max, float eta_max) : GenericBoat(radius, mass, inertia, mu_l, mu_ct,
+BoomBoat::BoomBoat(double radius, double mass, double inertia, double mu_l,
+ double mu_ct, double mu_r, Vector3d pos, Vector3d vel, double fuel, double cap,
+  double F_max, double eta_max) : GenericBoat(radius, mass, inertia, mu_l, mu_ct,
    mu_r, pos, vel, F_max, eta_max), fuel(fuel), tank_curr(fuel), cap(cap) {
-    this->set_control(Vector2f(0.0, 0.0));
+    this->set_control(Vector2d(0.0, 0.0));
     this->load_boom_boat_params("params.json");
 }
 
@@ -92,17 +92,17 @@ BoomBoat &BoomBoat::operator=(const BoomBoat &boom_boat) {
 }
 
 // Getter for fuel
-float BoomBoat::get_fuel() const { return fuel; }
+double BoomBoat::get_fuel() const { return fuel; }
 
 // Getter for cap
-float BoomBoat::get_cap() const { return cap; }
+double BoomBoat::get_cap() const { return cap; }
 
 
 // Getter for tank_curr
-float BoomBoat::get_tank_curr() const { return tank_curr; }
+double BoomBoat::get_tank_curr() const { return tank_curr; }
 
 // Setter for fuel
-void BoomBoat::set_fuel(float fuel) {
+void BoomBoat::set_fuel(double fuel) {
     if (fuel < 0) {
         std::cerr << "Fuel cannot be negative." << std::endl;
         return;
@@ -123,7 +123,7 @@ void BoomBoat::load_boom_boat_params(std::string filename) {
 }
 
 // Check if the control is valid
-// bool BoomBoat::is_valid_control(Vector2f control) const {
+// bool BoomBoat::is_valid_control(Vector2d control) const {
 //     // Check if the abs(force) is within the limits
 //     if (abs(control(0)) > this->get_F_max()) {
 //         std::cerr << "Force exceeds the maximum limit: " << this->get_F_max() << " [N]" << std::endl;
@@ -147,8 +147,8 @@ void BoomBoat::load_boom_boat_params(std::string filename) {
 //     json params;
 //     file >> params;
 
-//     float Lips_F = params["generic_boat"]["lipschitz_cont_F"];
-//     float Lips_eta = params["generic_boat"]["lipschitz_cont_eta_deg"];
+//     double Lips_F = params["generic_boat"]["lipschitz_cont_F"];
+//     double Lips_eta = params["generic_boat"]["lipschitz_cont_eta_deg"];
 //     Lips_eta = Lips_eta * PI / 180.0;
 
 //     // Check Lipschitz continuity
@@ -177,43 +177,45 @@ void BoomBoat::print_status() const {
 }
 
 // State derivative function
-VectorXf BoomBoat::state_der(VectorXf state, Vector2f control,
-     Vector2f boom_force) const {
+VectorXd BoomBoat::state_der(VectorXd state, Vector2d control,
+     Vector2d boom_force) const {
     // state = [x, y, theta, x_dot, y_dot, omega]
     // control = [F, eta]
 
     // Unpack state
-    // float x = state(0);
-    // float y = state(1);
-    float theta = state(2);
-    float x_dot = state(3);
-    float y_dot = state(4);
-    float omega = state(5);
+    // double x = state(0);
+    // double y = state(1);
+    double theta = state(2);
+    double x_dot = state(3);
+    double y_dot = state(4);
+    double omega = state(5);
     // Unpack control
-    float F = control(0);
-    float eta = control(1);
+    double F = control(0);
+    double eta = control(1);
 
     // Unpack boat parameters
-    float r = this->get_radius();
-    float mass = this->get_mass();
-    float I = this->get_inertia();
-    float mu_r = this->get_mu_r();
-    float mu_l = this->get_mu_l();
-    float mu_ct = this->get_mu_ct();
+    double r = this->get_radius();
+    double mass = this->get_mass();
+    double I = this->get_inertia();
+    double mu_r = this->get_mu_r();
+    double mu_l = this->get_mu_l();
+    double mu_ct = this->get_mu_ct();
 
     // Rotation to local frame
-    float u = x_dot * sin(theta) + y_dot * cos(theta);
-    float v = x_dot * cos(theta) - y_dot * sin(theta);
+    double u = x_dot * sin(theta) + y_dot * cos(theta);
+    double v = x_dot * cos(theta) - y_dot * sin(theta);
 
-    float F1 = boom_force(0);
-    float F2 = boom_force(1);
-    float F_boom_u = F1 * sin(theta) + F2 * cos(theta);
-    float F_boom_v = F1 * cos(theta) - F2 * sin(theta);
+    double F1 = boom_force(0);
+    double F2 = boom_force(1);
+    double F_boom_u = F1 * sin(theta) + F2 * cos(theta);
+    double F_boom_v = F1 * cos(theta) - F2 * sin(theta);
 
     // EOM
-    float u_dot = (F * cos(eta) - mu_l * u * u * sign(u) + F_boom_u) / mass;
-    float v_dot = (-F * sin(eta) - mu_ct * v * v * sign(v) + F_boom_v) / mass;
-    float omega_dot = (r * F * sin(eta) - mu_r * omega * omega * sign(omega) - r * F_boom_v) / I;
+    double u_dot = (F * cos(eta) - mu_l * u * u * sign(u) + F_boom_u) / mass;
+    double v_dot = (-F * sin(eta) - mu_ct * v * v * sign(v) + F_boom_v) / mass;
+    double omega_dot = (r * F * sin(eta) - mu_r * omega * omega * sign(omega) - r * F_boom_v) / I;
+    
+
 
     // // print u_dot
     // cout << "u_dot: " << u_dot << endl;
@@ -222,10 +224,10 @@ VectorXf BoomBoat::state_der(VectorXf state, Vector2f control,
     // cout.flush();
 
     // Rotation backward to global frame
-    float x_dotdot = u_dot * sin(theta) + v_dot * cos(theta);
-    float y_dotdot = u_dot * cos(theta) - v_dot * sin(theta);
+    double x_dotdot = u_dot * sin(theta) + v_dot * cos(theta);
+    double y_dotdot = u_dot * cos(theta) - v_dot * sin(theta);
 
-    VectorXf state_dot(6);
+    VectorXd state_dot(6);
     state_dot << x_dot, y_dot, omega, x_dotdot, y_dotdot, omega_dot;
     return state_dot;
     }
