+++ boom-boat.h
@@ -8,19 +8,19 @@
 #include "json/json.hpp"
 using json = nlohmann::json;
 
-using Eigen::Vector3f;
-using Eigen::Vector2f;
-using Eigen::VectorXf;
-using Eigen::MatrixXf;
+using Eigen::Vector3d;
+using Eigen::Vector2d;
+using Eigen::VectorXd;
+using Eigen::MatrixXd;
 
 class BoomBoat : public GenericBoat {
 private:
     // Member variables
-    float fuel; // Fuel [kg]
-    float tank_curr; // Waste oil in tank [L]
-    float cap; // Waste oil capacity [L]
-    // const float F_max; // Maximum force applied by engine [N]
-    // const float eta_max; // Maximum steering angle [rad]
+    double fuel; // Fuel [kg]
+    double tank_curr; // Waste oil in tank [L]
+    double cap; // Waste oil capacity [L]
+    // const double F_max; // Maximum force applied by engine [N]
+    // const double eta_max; // Maximum steering angle [rad]
 
     json boom_boat_params;
 
@@ -28,10 +28,10 @@ public:
     // Constructors
     BoomBoat(); // Default constructor
     BoomBoat(const BoomBoat &boom_boat); // Copy constructor
-    BoomBoat(Vector3f pos);
-    BoomBoat(float radius, float mass, float inertia, float mu_l, float mu_ct, 
-             float mu_r, Vector3f pos, Vector3f vel, float fuel, float cap,
-              float F_max, float eta_max); // Parameterized constructor
+    BoomBoat(Vector3d pos);
+    BoomBoat(double radius, double mass, double inertia, double mu_l, double mu_ct, 
+             double mu_r, Vector3d pos, Vector3d vel, double fuel, double cap,
+              double F_max, double eta_max); // Parameterized constructor
 
     // Destructor
     ~BoomBoat();
@@ -40,22 +40,22 @@ public:
     BoomBoat &operator=(const BoomBoat &boom_boat);
 
     // Getter and Setter for fuel
-    float get_fuel() const;
-    float get_cap() const;
-    float get_tank_curr() const;
+    double get_fuel() const;
+    double get_cap() const;
+    double get_tank_curr() const;
     
-    void set_fuel(float fuel);
+    void set_fuel(double fuel);
 
     void load_boom_boat_params(std::string filename);
 
-    // bool is_valid_control(Vector2f control) const;
+    // bool is_valid_control(Vector2d control) const;
     
     // Utility function to display BoomBoat status
     void print_status() const;
 
     // state derivative function
-    VectorXf state_der(VectorXf state, Vector2f control,
-     Vector2f boom_force) const;
+    VectorXd state_der(VectorXd state, Vector2d control,
+     Vector2d boom_force) const;
 };
 
 #endif // BOOM_BOAT_H
