+++ dubin.cpp
@@ -54,16 +54,16 @@
 // #define EDUBNOPATH    (4)   // no connection between configurations with this word
 
 // // The various types of solvers for each of the path types
-// typedef int (*DubinsWord)(float, float, float, float* );
+// typedef int (*DubinsWord)(double, double, double, double* );
 
 // // A complete list of the possible solvers that could give optimal paths
 // extern DubinsWord dubins_words[];
 
 // typedef struct
 // {
-//     float qi[3];       // the initial configuration
-//     float param[3];    // the lengths of the three segments
-//     float rho;         // model forward velocity / model angular velocity
+//     double qi[3];       // the initial configuration
+//     double param[3];    // the lengths of the three segments
+//     double rho;         // model forward velocity / model angular velocity
 //     int type;           // path type. one of LSL, LSR, ...
 // } DubinsPath;
 
@@ -75,7 +75,7 @@
 //  * @note the user_data parameter is forwarded from the caller
 //  * @note return non-zero to denote sampling should be stopped
 //  */
-// typedef int (*DubinsPathSamplingCallback)(float q[3], float t, void* user_data);
+// typedef int (*DubinsPathSamplingCallback)(double q[3], double t, void* user_data);
 
 // /**
 //  * Generate a path from an initial configuration to
@@ -91,14 +91,14 @@
 //  * @param path  - the resultant path
 //  * @return      - non-zero on error
 //  */
-// int dubins_init( float q0[3], float q1[3], float rho, DubinsPath* path);
+// int dubins_init( double q0[3], double q1[3], double rho, DubinsPath* path);
 
 // /**
 //  * Calculate the length of an initialised path
 //  *
 //  * @param path - the path to find the length of
 //  */
-// float dubins_path_length( DubinsPath* path );
+// double dubins_path_length( DubinsPath* path );
 
 // /**
 //  * Extract an integer that represents which path type was used
@@ -116,7 +116,7 @@
 //  * @param q    - the configuration result
 //  * @returns    - non-zero if 't' is not in the correct range
 //  */
-// int dubins_path_sample( DubinsPath* path, float t, float q[3]);
+// int dubins_path_sample( DubinsPath* path, double t, double q[3]);
 
 // /**
 //  * Walk along the path at a fixed sampling interval, calling the
@@ -127,7 +127,7 @@
 //  * @param user_data - optional information to pass on to the callback
 //  * @param stepSize  - the distance along the path for subsequent samples
 //  */
-// int dubins_path_sample_many( DubinsPath* path, DubinsPathSamplingCallback cb, float stepSize, void* user_data );
+// int dubins_path_sample_many( DubinsPath* path, DubinsPathSamplingCallback cb, double stepSize, void* user_data );
 
 // /**
 //  * Convenience function to identify the endpoint of a path
@@ -135,7 +135,7 @@
 //  * @param path - an initialised path
 //  * @param q    - the configuration result
 //  */
-// int dubins_path_endpoint( DubinsPath* path, float q[3] );
+// int dubins_path_endpoint( DubinsPath* path, double q[3] );
 
 // /**
 //  * Convenience function to extract a subset of a path
@@ -144,15 +144,15 @@
 //  * @param t       - a length measure, where 0 < t < dubins_path_length(path)
 //  * @param newpath - the resultant path
 //  */
-// int dubins_extract_subpath( DubinsPath* path, float t, DubinsPath* newpath );
+// int dubins_extract_subpath( DubinsPath* path, double t, DubinsPath* newpath );
 
 // // Only exposed for testing purposes
-// int dubins_LSL( float alpha, float beta, float d, float* outputs );
-// int dubins_RSR( float alpha, float beta, float d, float* outputs );
-// int dubins_LSR( float alpha, float beta, float d, float* outputs );
-// int dubins_RSL( float alpha, float beta, float d, float* outputs );
-// int dubins_LRL( float alpha, float beta, float d, float* outputs );
-// int dubins_RLR( float alpha, float beta, float d, float* outputs );
+// int dubins_LSL( double alpha, double beta, double d, double* outputs );
+// int dubins_RSR( double alpha, double beta, double d, double* outputs );
+// int dubins_LSR( double alpha, double beta, double d, double* outputs );
+// int dubins_RSL( double alpha, double beta, double d, double* outputs );
+// int dubins_LRL( double alpha, double beta, double d, double* outputs );
+// int dubins_RLR( double alpha, double beta, double d, double* outputs );
 
 
 // The segment types for each of the Path types
@@ -175,11 +175,11 @@ DubinsWord dubins_words[] = {
 };
 
 #define UNPACK_INPUTS(alpha, beta)     \
-    float sa = sin(alpha);            \
-    float sb = sin(beta);             \
-    float ca = cos(alpha);            \
-    float cb = cos(beta);             \
-    float c_ab = cos(alpha - beta);   \
+    double sa = sin(alpha);            \
+    double sb = sin(beta);             \
+    double ca = cos(alpha);            \
+    double cb = cos(beta);             \
+    double c_ab = cos(alpha - beta);   \
 
 #define PACK_OUTPUTS(outputs)       \
     outputs[0]  = t;                \
@@ -191,28 +191,28 @@ DubinsWord dubins_words[] = {
  *
  * fmod doesn't behave correctly for angular quantities, this function does
  */
-float fmodr( float x, float y)
+double fmodr( double x, double y)
 {
     return x - y*floor(x/y);
 }
 
-float mod2pi( float theta )
+double mod2pi( double theta )
 {
     return fmodr( theta, 2 * M_PI );
 }
 
-int dubins_init_normalised( float alpha, float beta, float d, DubinsPath* path)
+int dubins_init_normalised( double alpha, double beta, double d, DubinsPath* path)
 {
-    float best_cost = INFINITY;
+    double best_cost = INFINITY;
     int    best_word;
     int    i;
 
     best_word = -1;
     for( i = 0; i < 6; i++ ) {
-        float params[3];
+        double params[3];
         int err = dubins_words[i](alpha, beta, d, params);
         if(err == EDUBOK) {
-            float cost = params[0] + params[1] + params[2];
+            double cost = params[0] + params[1] + params[2];
             if(cost < best_cost) {
                 best_word = i;
                 best_cost = cost;
@@ -231,19 +231,19 @@ int dubins_init_normalised( float alpha, float beta, float d, DubinsPath* path)
     return EDUBOK;
 }
 
-int dubins_init( float q0[3], float q1[3], float rho, DubinsPath* path )
+int dubins_init( double q0[3], double q1[3], double rho, DubinsPath* path )
 {
     int i;
-    float dx = q1[0] - q0[0];
-    float dy = q1[1] - q0[1];
-    float D = sqrt( dx * dx + dy * dy );
-    float d = D / rho;
+    double dx = q1[0] - q0[0];
+    double dy = q1[1] - q0[1];
+    double D = sqrt( dx * dx + dy * dy );
+    double d = D / rho;
     if( rho <= 0. ) {
         return EDUBBADRHO;
     }
-    float theta = mod2pi(atan2( dy, dx ));
-    float alpha = mod2pi(q0[2] - theta);
-    float beta  = mod2pi(q1[2] - theta);
+    double theta = mod2pi(atan2( dy, dx ));
+    double alpha = mod2pi(q0[2] - theta);
+    double beta  = mod2pi(q1[2] - theta);
     for( i = 0; i < 3; i ++ ) {
         path->qi[i] = q0[i];
     }
@@ -252,99 +252,99 @@ int dubins_init( float q0[3], float q1[3], float rho, DubinsPath* path )
     return dubins_init_normalised( alpha, beta, d, path );
 }
 
-int dubins_LSL( float alpha, float beta, float d, float* outputs )
+int dubins_LSL( double alpha, double beta, double d, double* outputs )
 {
     UNPACK_INPUTS(alpha, beta);
-    float tmp0 = d+sa-sb;
-    float p_squared = 2 + (d*d) -(2*c_ab) + (2*d*(sa - sb));
+    double tmp0 = d+sa-sb;
+    double p_squared = 2 + (d*d) -(2*c_ab) + (2*d*(sa - sb));
     if( p_squared < 0 ) {
         return EDUBNOPATH;
     }
-    float tmp1 = atan2( (cb-ca), tmp0 );
-    float t = mod2pi(-alpha + tmp1 );
-    float p = sqrt( p_squared );
-    float q = mod2pi(beta - tmp1 );
+    double tmp1 = atan2( (cb-ca), tmp0 );
+    double t = mod2pi(-alpha + tmp1 );
+    double p = sqrt( p_squared );
+    double q = mod2pi(beta - tmp1 );
     PACK_OUTPUTS(outputs);
     return EDUBOK;
 }
 
-int dubins_RSR( float alpha, float beta, float d, float* outputs )
+int dubins_RSR( double alpha, double beta, double d, double* outputs )
 {
     UNPACK_INPUTS(alpha, beta);
-    float tmp0 = d-sa+sb;
-    float p_squared = 2 + (d*d) -(2*c_ab) + (2*d*(sb-sa));
+    double tmp0 = d-sa+sb;
+    double p_squared = 2 + (d*d) -(2*c_ab) + (2*d*(sb-sa));
     if( p_squared < 0 ) {
         return EDUBNOPATH;
     }
-    float tmp1 = atan2( (ca-cb), tmp0 );
-    float t = mod2pi( alpha - tmp1 );
-    float p = sqrt( p_squared );
-    float q = mod2pi( -beta + tmp1 );
+    double tmp1 = atan2( (ca-cb), tmp0 );
+    double t = mod2pi( alpha - tmp1 );
+    double p = sqrt( p_squared );
+    double q = mod2pi( -beta + tmp1 );
     PACK_OUTPUTS(outputs);
     return EDUBOK;
 }
 
-int dubins_LSR( float alpha, float beta, float d, float* outputs )
+int dubins_LSR( double alpha, double beta, double d, double* outputs )
 {
     UNPACK_INPUTS(alpha, beta);
-    float p_squared = -2 + (d*d) + (2*c_ab) + (2*d*(sa+sb));
+    double p_squared = -2 + (d*d) + (2*c_ab) + (2*d*(sa+sb));
     if( p_squared < 0 ) {
         return EDUBNOPATH;
     }
-    float p    = sqrt( p_squared );
-    float tmp2 = atan2( (-ca-cb), (d+sa+sb) ) - atan2(-2.0, p);
-    float t    = mod2pi(-alpha + tmp2);
-    float q    = mod2pi( -mod2pi(beta) + tmp2 );
+    double p    = sqrt( p_squared );
+    double tmp2 = atan2( (-ca-cb), (d+sa+sb) ) - atan2(-2.0, p);
+    double t    = mod2pi(-alpha + tmp2);
+    double q    = mod2pi( -mod2pi(beta) + tmp2 );
     PACK_OUTPUTS(outputs);
     return EDUBOK;
 }
 
-int dubins_RSL( float alpha, float beta, float d, float* outputs )
+int dubins_RSL( double alpha, double beta, double d, double* outputs )
 {
     UNPACK_INPUTS(alpha, beta);
-    float p_squared = (d*d) -2 + (2*c_ab) - (2*d*(sa+sb));
+    double p_squared = (d*d) -2 + (2*c_ab) - (2*d*(sa+sb));
     if( p_squared< 0 ) {
         return EDUBNOPATH;
     }
-    float p    = sqrt( p_squared );
-    float tmp2 = atan2( (ca+cb), (d-sa-sb) ) - atan2(2.0, p);
-    float t    = mod2pi(alpha - tmp2);
-    float q    = mod2pi(beta - tmp2);
+    double p    = sqrt( p_squared );
+    double tmp2 = atan2( (ca+cb), (d-sa-sb) ) - atan2(2.0, p);
+    double t    = mod2pi(alpha - tmp2);
+    double q    = mod2pi(beta - tmp2);
     PACK_OUTPUTS(outputs);
     return EDUBOK;
 }
 
-int dubins_RLR( float alpha, float beta, float d, float* outputs )
+int dubins_RLR( double alpha, double beta, double d, double* outputs )
 {
     UNPACK_INPUTS(alpha, beta);
-    float tmp_rlr = (6. - d*d + 2*c_ab + 2*d*(sa-sb)) / 8.;
+    double tmp_rlr = (6. - d*d + 2*c_ab + 2*d*(sa-sb)) / 8.;
     if( fabs(tmp_rlr) > 1) {
         return EDUBNOPATH;
     }
-    float p = mod2pi( 2*M_PI - acos( tmp_rlr ) );
-    float t = mod2pi(alpha - atan2( ca-cb, d-sa+sb ) + mod2pi(p/2.));
-    float q = mod2pi(alpha - beta - t + mod2pi(p));
+    double p = mod2pi( 2*M_PI - acos( tmp_rlr ) );
+    double t = mod2pi(alpha - atan2( ca-cb, d-sa+sb ) + mod2pi(p/2.));
+    double q = mod2pi(alpha - beta - t + mod2pi(p));
     PACK_OUTPUTS( outputs );
     return EDUBOK;
 }
 
-int dubins_LRL( float alpha, float beta, float d, float* outputs )
+int dubins_LRL( double alpha, double beta, double d, double* outputs )
 {
     UNPACK_INPUTS(alpha, beta);
-    float tmp_lrl = (6. - d*d + 2*c_ab + 2*d*(- sa + sb)) / 8.;
+    double tmp_lrl = (6. - d*d + 2*c_ab + 2*d*(- sa + sb)) / 8.;
     if( fabs(tmp_lrl) > 1) {
         return EDUBNOPATH;
     }
-    float p = mod2pi( 2*M_PI - acos( tmp_lrl ) );
-    float t = mod2pi(-alpha - atan2( ca-cb, d+sa-sb ) + p/2.);
-    float q = mod2pi(mod2pi(beta) - alpha -t + mod2pi(p));
+    double p = mod2pi( 2*M_PI - acos( tmp_lrl ) );
+    double t = mod2pi(-alpha - atan2( ca-cb, d+sa-sb ) + p/2.);
+    double q = mod2pi(mod2pi(beta) - alpha -t + mod2pi(p));
     PACK_OUTPUTS( outputs );
     return EDUBOK;
 }
 
-float dubins_path_length( DubinsPath* path )
+double dubins_path_length( DubinsPath* path )
 {
-    float length = 0.;
+    double length = 0.;
     length += path->param[0];
     length += path->param[1];
     length += path->param[2];
@@ -356,7 +356,7 @@ int dubins_path_type( DubinsPath* path ) {
     return path->type;
 }
 
-void dubins_segment( float t, float qi[3], float qt[3], int type)
+void dubins_segment( double t, double qi[3], double qt[3], int type)
 {
     assert( type == L_SEG || type == S_SEG || type == R_SEG );
 
@@ -377,7 +377,7 @@ void dubins_segment( float t, float qi[3], float qt[3], int type)
     }
 }
 
-int dubins_path_sample( DubinsPath* path, float t, float q[3] )
+int dubins_path_sample( DubinsPath* path, double t, double q[3] )
 {
     if( t < 0 || t >= dubins_path_length(path) ) {
         // error, parameter out of bounds
@@ -385,7 +385,7 @@ int dubins_path_sample( DubinsPath* path, float t, float q[3] )
     }
 
     // tprime is the normalised variant of the parameter t
-    float tprime = t / path->rho;
+    double tprime = t / path->rho;
 
     // In order to take rho != 1 into account this function needs to be more complex
     // than it would be otherwise. The transformation is done in five stages.
@@ -398,14 +398,14 @@ int dubins_path_sample( DubinsPath* path, float t, float q[3] )
     //      normalise the target configurations angular component
 
     // The translated initial configuration
-    float qi[3] = { 0, 0, path->qi[2] };
+    double qi[3] = { 0, 0, path->qi[2] };
 
     // Generate the target configuration
     const int* types = DIRDATA[path->type];
-    float p1 = path->param[0];
-    float p2 = path->param[1];
-    float q1[3]; // end-of segment 1
-    float q2[3]; // end-of segment 2
+    double p1 = path->param[0];
+    double p2 = path->param[1];
+    double q1[3]; // end-of segment 1
+    double q2[3]; // end-of segment 2
     dubins_segment( p1,      qi,    q1, types[0] );
     dubins_segment( p2,      q1,    q2, types[1] );
     if( tprime < p1 ) {
@@ -426,12 +426,12 @@ int dubins_path_sample( DubinsPath* path, float t, float q[3] )
     return 0;
 }
 
-int dubins_path_sample_many( DubinsPath* path, DubinsPathSamplingCallback cb, float stepSize, void* user_data )
+int dubins_path_sample_many( DubinsPath* path, DubinsPathSamplingCallback cb, double stepSize, void* user_data )
 {
-    float x = 0.0;
-    float length = dubins_path_length(path);
+    double x = 0.0;
+    double length = dubins_path_length(path);
     while( x <  length ) {
-        float q[3];
+        double q[3];
         dubins_path_sample( path, x, q );
         int retcode = cb(q, x, user_data);
         if( retcode != 0 ) {
@@ -442,16 +442,16 @@ int dubins_path_sample_many( DubinsPath* path, DubinsPathSamplingCallback cb, fl
     return 0;
 }
 
-int dubins_path_endpoint( DubinsPath* path, float q[3] )
+int dubins_path_endpoint( DubinsPath* path, double q[3] )
 {
     // TODO - introduce a new constant rather than just using EPSILON
     return dubins_path_sample( path, dubins_path_length(path) - EPSILON, q );
 }
 
-int dubins_extract_subpath( DubinsPath* path, float t, DubinsPath* newpath )
+int dubins_extract_subpath( DubinsPath* path, double t, DubinsPath* newpath )
 {
     // calculate the true parameter
-    float tprime = t / path->rho;
+    double tprime = t / path->rho;
 
     // copy most of the data
     newpath->qi[0] = path->qi[0];
@@ -467,7 +467,7 @@ int dubins_extract_subpath( DubinsPath* path, float t, DubinsPath* newpath )
     return 0;
 }
 
-void save_to_file(string file_name, string folder_name, MatrixXf path, MatrixXf path_R, MatrixXf path_L) {
+void save_to_file(string file_name, string folder_name, MatrixXd path, MatrixXd path_R, MatrixXd path_L) {
     string folder_path = folder_name + "/";
     string file_path = folder_path + file_name + ".txt";
     std::ofstream file(file_path);
